LayoutTests/media/video-play-empty-events-expected.txt

Test that play() from EMPTY network state triggers load() and async play event.

EXPECTED (video.networkState == '0') OK
RUN(video.play())
EVENT(loadstart)
EXPECTED (video.paused == 'true') OK
SCRIPT DONE
EVENT(play)
EXPECTED (video.paused == 'false') OK
END OF TEST
